+++ models/intermediate/int_iterable__list_user_unnest.sql
@@ -1,27 +1,64 @@
-{{ config( materialized='table') }}
--- materializing as a table because the computations here are fairly complex
+{{ config(
+        materialized='incremental',
+        unique_key='unique_key',
+        incremental_strategy='insert_overwrite' if target.type in ('bigquery', 'spark', 'databricks') else 'delete+insert',
+        partition_by={"field": "date_day", "data_type": "date"} if target.type not in ('spark','databricks') else ['date_day'],
+        file_format='delta',
+        on_schema_change='fail'
+    ) 
+}}
 
 with user_history as (
 
     select * 
-    from {{ ref('int_iterable__list_user_history') }}
+    from {{ ref('int_iterable__list_user_history') }} as user_history
 
-/*
-Unfortunately, `email_list_ids` are brought in as a JSON ARRAY, which different destinations handle completely differently. 
-The below code serves to extract and pivot list IDs into individual rows. 
-Records with an empty `email_list_ids` array will just have one row. 
-We are making the assumption that a user will not have more than 1000 lists. If that's wrong please open an issue!
-https://github.com/fivetran/dbt_iterable/issues/new/choose
-*/
+    {% if is_incremental() %}
+    {# the only rows we potentially want to overwrite are  active ones  #}
+    where user_history.updated_at >= coalesce((select min(updated_at) from {{ this }} where is_current), '2010-01-01')
+    {% endif %}
+
 {% if target.type == 'redshift' %}
-), numbers as (
-    select 0 as generated_number
-    union 
-    select *
-    from (
-        {{ dbt_utils.generate_series(upper_bound=1000) }} )
-{% endif %}
+{# using PartiQL syntax to work with redshift's SUPER types #}
+), redshift_parse_email_lists as (
+
+    select
+        email,
+        first_name,
+        last_name,
+        user_id,
+        signup_date,
+        signup_source,
+        phone_number,
+        updated_at,
+        is_current,
+        email_list_ids,
+        {# let's not remove empty array-rows #}
+        json_parse(case when email_list_ids = '[]' then '["is_null"]' else email_list_ids end) as super_email_list_ids
+        
+    from user_history
+
 ), unnest_email_array as (
+
+    select
+        email,
+        first_name,
+        last_name,
+        user_id,
+        signup_date,
+        signup_source,
+        phone_number,
+        updated_at,
+        is_current,
+        {# go back to strings #}
+        cast(email_list_ids as {{ dbt.type_string() }}) as email_list_ids, 
+        cast(email_list_id as {{ dbt.type_string() }}) as email_list_id
+
+    from redshift_parse_email_lists as emails, emails.super_email_list_ids as email_list_id
+
+{% else %}
+), unnest_email_array as (
+
     select
         email,
         first_name,
@@ -36,33 +73,44 @@ https://github.com/fivetran/dbt_iterable/issues/new/choose
         case when email_list_ids != '[]' then
         {% if target.type == 'snowflake' %}
         email_list_id.value
-        {% elif target.type == 'redshift' %}
-        json_extract_array_element_text(email_list_ids, cast(numbers.generated_number as {{ dbt.type_int() }}), true) 
-        {% else %} email_list_id
-        {% endif %}
-        else null end 
-        as 
-        email_list_id
+        {% else %} email_list_id {% endif %} else null end as email_list_id
 
     from user_history
 
-    cross join 
     {% if target.type == 'snowflake' %}
-        table(flatten(cast(email_list_ids as VARIANT))) as email_list_id 
+    cross join 
+        table(flatten(input => parse_json(email_list_ids))) as email_list_id
     {% elif target.type == 'bigquery' %}
+    cross join 
         unnest(JSON_EXTRACT_STRING_ARRAY(email_list_ids)) as email_list_id
-    {% elif target.type == 'redshift' %}
-        numbers 
-    where numbers.generated_number < json_array_length(email_list_ids, true)
-        or (numbers.generated_number + json_array_length(email_list_ids, true) = 0)
     {% else %}
-    /* postgres */
+    {# postgres #}
+    cross join 
         json_array_elements_text(cast((
-            case when email_list_ids = '[]' then '["is_null"]'  -- to not remove empty array-rows
+            case when email_list_ids = '[]' then '["is_null"]' {# to not remove empty array-rows #}
             else email_list_ids end) as json)) as email_list_id
     {%- endif %}
 
+{%- endif -%}
+), adjust_nulls as (
+
+    select
+        email,
+        first_name,
+        last_name,
+        user_id,
+        signup_date,
+        signup_source,
+        updated_at,
+        phone_number,
+        is_current,
+        case when email_list_ids = '["is_null"]' then '[]' else email_list_ids end as email_list_ids,
+        cast(NULLIF(email_list_id, 'is_null') as {{ dbt.type_int() }}) as list_id
+
+    from unnest_email_array
+
 ), final as (
+
     select
         email,
         first_name,
@@ -74,8 +122,11 @@ https://github.com/fivetran/dbt_iterable/issues/new/choose
         phone_number,
         is_current,
         email_list_ids,
-        cast(email_list_id as {{ dbt.type_int() }}) as list_id
-    from unnest_email_array
+        list_id,
+        {{ dbt_utils.generate_surrogate_key(["email", "list_id", "updated_at"]) }} as unique_key,
+        cast( {{ dbt.date_trunc('day', 'updated_at') }} as date) as date_day
+    
+    from adjust_nulls
 )
 
 select *